+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added a dialog to get people to add a future dependency. Since Rimworld doesn't have a good mechanism to automatically add new dependencies, we need to be a little in-your-face about it.
 -   Rimworld 1.6 support.
 
+### Changed
+
+-    Attempted to make the transpiler for the Dynamic Trade Interface integration more robust
+
 ## [0.8.1] - 2024-07-06
 
 ### Changed